Update lewagon-cecile.txt
I have checked the wording, which is correct.  However, are you going to publish it as it is on your blog?  If so, I would suggest editing it a little so that it reads more like a magazine interview - i.e. taking out the various "like" and "umm" .  Let me know as I can do this really easily.  Also, as a blog post I would not out the initial bit as this is a repeat as to what is below.  What about a small intro such as "In this blog post, Eddie Jaoude interviews Cecile [include surname] on her journey to becoming a developer through the Le Wagon Bootcamp, with comments from [insert name of speaker 3] from Le Wagon." I think this will make it more engaging to the reader.